requirements.txt
streamlit
python-dotenv
openai
sentence-transformers
torch
chromadb
pdfplumber
neo4j

